+++ src/Container/HandleMessageMiddlewareFactory.php
@@ -4,16 +4,29 @@ declare(strict_types=1);
 
 namespace Xtreamwayz\Expressive\Messenger\Container;
 
+use InvalidArgumentException;
 use Psr\Container\ContainerInterface;
 use Symfony\Component\Messenger\Middleware\HandleMessageMiddleware;
 use Symfony\Component\Messenger\Middleware\MiddlewareInterface;
+use function sprintf;
 
 class HandleMessageMiddlewareFactory
 {
     /** @var string */
     private $busName;
 
-    public function __construct(string $busName = 'messenger.bus.default')
+    public static function __callStatic(string $busName, array $arguments) : MiddlewareInterface
+    {
+        if (! isset($arguments[0]) || ! $arguments[0] instanceof ContainerInterface) {
+            throw new InvalidArgumentException(
+                sprintf('The first argument must be of type %s', ContainerInterface::class)
+            );
+        }
+
+        return (new self($busName))($arguments[0]);
+    }
+
+    public function __construct(string $busName = 'messenger.default.bus')
     {
         $this->busName = $busName;
     }